create 404 \ add Router \ create private route
