+++ src/components/layout/Layout.tsx
@@ -1,11 +1,13 @@
-function FavoritesEmpty(): JSX.Element {
+import { Outlet, Link } from 'react-router-dom';
+
+function Layout(): JSX.Element {
   return (
-    <div className="page page--favorites-empty">
+    <div className="page page--gray page--main">
       <header className="header">
         <div className="container">
           <div className="header__wrapper">
             <div className="header__left">
-              <a className="header__logo-link" href="main.html">
+              <Link to="/" className="header__logo-link header__logo-link--active">
                 <img
                   className="header__logo"
                   src="img/logo.svg"
@@ -13,7 +15,7 @@ function FavoritesEmpty(): JSX.Element {
                   width={81}
                   height={41}
                 />
-              </a>
+              </Link>
             </div>
             <nav className="header__nav">
               <ul className="header__nav-list">
@@ -26,7 +28,7 @@ function FavoritesEmpty(): JSX.Element {
                     <span className="header__user-name user__name">
                       Oliver.conner@gmail.com
                     </span>
-                    <span className="header__favorite-count">0</span>
+                    <span className="header__favorite-count">3</span>
                   </a>
                 </li>
                 <li className="header__nav-item">
@@ -39,32 +41,9 @@ function FavoritesEmpty(): JSX.Element {
           </div>
         </div>
       </header>
-      <main className="page__main page__main--favorites page__main--favorites-empty">
-        <div className="page__favorites-container container">
-          <section className="favorites favorites--empty">
-            <h1 className="visually-hidden">Favorites (empty)</h1>
-            <div className="favorites__status-wrapper">
-              <b className="favorites__status">Nothing yet saved.</b>
-              <p className="favorites__status-description">
-                Save properties to narrow down search or plan your future trips.
-              </p>
-            </div>
-          </section>
-        </div>
-      </main>
-      <footer className="footer">
-        <a className="footer__logo-link" href="main.html">
-          <img
-            className="footer__logo"
-            src="img/logo.svg"
-            alt="6 cities logo"
-            width={64}
-            height={33}
-          />
-        </a>
-      </footer>
+      <Outlet />
     </div>
   );
 }
 
-export default FavoritesEmpty;
+export default Layout;
