+++ src/components/pages/page-favorites-empty/FavoritesEmpty.tsx
@@ -1,0 +1,34 @@
+function FavoritesEmpty(): JSX.Element {
+  return (
+    <>
+      <main className="page__main page__main--favorites page__main--favorites-empty">
+        <div className="page__favorites-container container">
+          <section className="favorites favorites--empty">
+            <h1 className="visually-hidden">Favorites (empty)</h1>
+            <div className="favorites__status-wrapper">
+              <b className="favorites__status">Nothing yet saved.</b>
+              <p className="favorites__status-description">
+                Save properties to narrow down search or plan your future trips.
+              </p>
+            </div>
+          </section>
+        </div>
+      </main>
+      <footer className="footer">
+        {/* link to main */}
+        <a className="footer__logo-link" href="#">
+          <img
+            className="footer__logo"
+            src="img/logo.svg"
+            alt="6 cities logo"
+            width={64}
+            height={33}
+          />
+        </a>
+        {/* link to main */}
+      </footer>
+    </>
+  );
+}
+
+export default FavoritesEmpty;
